Even better memory

+++ stitcher.py
@@ -147,9 +147,9 @@ def blend_images(imgA, imgB, overlap_width, direction='horizontal'):
         overlapB = imgB[:, :overlap_width]
         
         alpha = np.linspace(1, 0, overlap_width).reshape(1, -1, 1)
-        blended_region = alpha * overlapA + (1 - alpha) * overlapB
+        blended_region = (alpha * overlapA + (1 - alpha) * overlapB).astype(np.uint8)
         
-        stitched = np.concatenate([imgA[:, :-overlap_width], blended_region, imgB[:, overlap_width:]], axis=1).astype(np.uint8)
+        stitched = np.concatenate([imgA[:, :-overlap_width], blended_region, imgB[:, overlap_width:]], axis=1)
         return stitched
     else:
         # Vertical blending
@@ -157,9 +157,9 @@ def blend_images(imgA, imgB, overlap_width, direction='horizontal'):
         overlapB = imgB[:overlap_width, :]
         
         alpha = np.linspace(1, 0, overlap_width).reshape(-1, 1, 1)
-        blended_region = alpha * overlapA + (1 - alpha) * overlapB
+        blended_region = (alpha * overlapA + (1 - alpha) * overlapB).astype(np.uint8)
         
-        stitched = np.concatenate([imgA[:-overlap_width, :], blended_region, imgB[overlap_width:, :]], axis=0).astype(np.uint8)
+        stitched = np.concatenate([imgA[:-overlap_width, :], blended_region, imgB[overlap_width:, :]], axis=0)
         return stitched
 
 def main(images, vert_clip_fraction: float, horz_clip_fraction: float, output_dir: str, write_intermediates: bool = False):
